+++ gegelatilib/include/learn/classificationLearningAgent.h
@@ -80,10 +80,8 @@ namespace Learn {
 		* \param[in] iSet Set of Instruction used to compose Programs in the
 		*            learning process.
 		* \param[in] p The LearningParameters for the LearningAgent.
-		* \param[in] nbRegs The number of registers for the execution
-		*                   environment of Program.
 		*/
-		ClassificationLearningAgent(ClassificationLearningEnvironment& le, const Instructions::Set& iSet, const LearningParameters& p, const unsigned int nbRegs = 8) : BaseLearningAgent(le, iSet, p, nbRegs) {};
+		ClassificationLearningAgent(ClassificationLearningEnvironment& le, const Instructions::Set& iSet, const LearningParameters& p) : BaseLearningAgent(le, iSet, p) {};
 
 		/**
 		* \brief Specialization of the evaluateRoot method for classification purposes.
@@ -254,7 +252,7 @@ namespace Learn {
 		auto allRoots = this->tpg.getRootVertices();
 		auto& tpgRef = this->tpg;
 		auto& resultsPerRootRef = this->resultsPerRoot;
-		std::for_each(allRoots.begin(), allRoots.end(), [&rootsToKeep, &tpgRef, &resultsPerRootRef](const TPG::TPGVertex* vert)
+		std::for_each(allRoots.begin(), allRoots.end(), [&rootsToKeep, &tpgRef, &resultsPerRootRef, &results](const TPG::TPGVertex* vert)
 			{
 				const std::type_info& vertexType = typeid(*vert);
 				// Do not remove actions
@@ -264,6 +262,16 @@ namespace Learn {
 
 					// Keep only results of non-decimated roots.
 					resultsPerRootRef.erase(vert);
+
+					// Update results also
+					std::multimap < std::shared_ptr<EvaluationResult>, const TPG::TPGVertex*>::iterator iter = results.begin();
+					while (iter != results.end()) {
+						if (iter->second == vert) {
+							results.erase(iter);
+							break;
+						}
+						iter++;
+					}
 				}
 			});
 	}