+++ test/classificationLearningAgentTest.cpp
@@ -139,8 +139,9 @@ TEST_F(ClassificationLearningAgentTest, DecimateWorstRoots) {
 	params.mutation.tpg.maxInitOutgoingEdges = 2;
 	params.ratioDeletedRoots = 0.50;
 	params.mutation.tpg.nbRoots = 50; // Param used in decimation
+	params.nbThreads = 4;
 
-	Learn::ClassificationLearningAgent cla(fle, set, params, 4);
+	Learn::ClassificationLearningAgent cla(fle, set, params);
 
 	// Initialize and populate the TPG
 	cla.init(0);